pyproject.toml
[project]
name = "prompt-demo"
version = "0.1.0"
description = "LangChain and Gemini API demonstrations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.7",
]
